Consolidate connect logic into one file
