+++ src/components/Provider.tsx
@@ -12,8 +12,8 @@ export interface ProviderProps<A extends Action = AnyAction> {
   store: Store<FixTypeLater, A>
   /**
    * Optional context to be used internally in react-redux. Use React.createContext() to create a context to be used.
-   * If this is used, generate own connect HOC by using connectAdvanced, supplying the same context provided to the
-   * Provider. Initial value doesn't matter, as it is overwritten with the internal state of Provider.
+   * If this is used, you'll need to customize `connect` by supplying the same context provided to the Provider.
+   * Initial value doesn't matter, as it is overwritten with the internal state of Provider.
    */
   context?: Context<ReactReduxContextValue>
   children: ReactNode
