+++ src/exports.ts
@@ -1,6 +1,9 @@
 import Provider from './components/Provider'
 import type { ProviderProps } from './components/Provider'
-import type { ConnectProps } from './components/connectAdvanced'
+import connect, {
+  ConnectProps,
+  ConnectedProps,
+} from './components/connectAdvanced'
 import type {
   SelectorFactory,
   Selector,
@@ -16,7 +19,6 @@ import type {
 } from './connect/selectorFactory'
 import { ReactReduxContext } from './components/Context'
 import type { ReactReduxContextValue } from './components/Context'
-import connect, { ConnectedProps } from './connect/connect'
 
 import { useDispatch, createDispatchHook } from './hooks/useDispatch'
 import { useSelector, createSelectorHook } from './hooks/useSelector'
